worker-python/requirements.txt
paho-mqtt==1.6.1
grpcio==1.60.0
grpcio-tools==1.60.0
google-generativeai==0.3.2
python-dotenv==1.0.0
